py/microservices/admin/requirements.txt
Django==3.1.3
djangorestframework==3.12.2
mysqlclient==2.0.1
django-mysql==3.9
django-cors-headers==3.7.0
pika==1.2.0
python-decouple==3.4